Twitter spam scanner
Grabbing the latest tweets from a list of usernames and storing them in GCP Datastore
Grabbing any replies to any tweets which haven't been checked and are older than a half hour (i.e. give everything half an hour to heat up before we grab replies)
Running all unprocessed tweets against a spam classifier (currently: def is_spam(tweet): return False  but that'll change tomorrow)
I'm set up to send stuff to Slack, but as it's an extra app I can't actually install it in this workspace.

Currently broken: ensuring that a given reply is actually a reply to the parent tweet.